Moved some utility modules into utils, created a projects directory to house the project functions/constants (and allow for expansion) and rewrote/reordered require stantements to reflect new paths.
