+++ src/projects/constants.js
@@ -11,4 +11,3 @@ module.exports.PROJECT_TYPE_DESCRIPTIONS = {
   [BASIC_STORY]: 'a vanilla JS app which runs inside a News story.'
 };
 module.exports.CONFIG_FILE_NAME = 'aunty.config.js';
-module.exports.DEFAULT_PORT = process.env.PORT || 8000;
