+++ src/utils/logo.js
@@ -1,6 +1,6 @@
 // Ours
-const {bad, blue, cyan, dim, green, hvy, magenta, red, yellow} = require('./string-styles');
-const {NEWLINE, zipTemplateLiterals} = require('./utils/strings');
+const {bad, blue, cyan, dim, green, hvy, magenta, red, yellow} = require('./color');
+const {NEWLINE, zipTemplateLiterals} = require('./strings');
 
 const COLORS = [blue, cyan, green, magenta, red, yellow];
